requirements.txt
asgiref==3.2.7
Django==3.0.6
django-cors-headers==3.2.1
django-filter==2.2.0
django-python3-ldap==0.11.3
djangorestframework==3.11.0
gunicorn==20.0.4
ldap3==2.7
Markdown==3.2.1
mysqlclient==1.4.6
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
